+++ packages/manager/apps/dedicated/client/app/components/dedicated-cloud/datacenter/vmware-vdc-add/dedicatedCloud-vmware-vdc-add.component.js
@@ -1,0 +1,13 @@
+import controller from './dedicatedCloud-vmware-vdc-add.controller';
+import template from './dedicatedCloud-vmware-vdc-add.html';
+
+export default {
+  bindings: {
+    orderSapHana: '<',
+    goBack: '<',
+    serviceName: '<',
+    trackClick: '<',
+  },
+  controller,
+  template,
+};