+++ packages/manager/apps/dedicated/client/app/dedicated/nas/dashboard/nas.constants.js
@@ -1,5 +1,0 @@
-export const NASHA_URL = '#/nasha';
-
-export default {
-  NASHA_URL,
-};