+++ packages/manager/core/generator/app/templates/src/components/Error/Error.tsx
@@ -1,28 +1,11 @@
 import React, { useEffect } from 'react';
-import { useNavigate, useLocation } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-import { OsdsMessage } from '@ovhcloud/ods-components/message/react';
-import { OsdsText } from '@ovhcloud/ods-components/text/react';
-import { OsdsButton } from '@ovhcloud/ods-components/button/react';
-import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
-import { ODS_MESSAGE_TYPE } from '@ovhcloud/ods-components/message';
-import { ODS_TEXT_LEVEL, ODS_TEXT_SIZE } from '@ovhcloud/ods-components/text';
-import { ODS_BUTTON_VARIANT } from '@ovhcloud/ods-components/button';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { useShell } from '@ovh-ux/manager-react-core-application';
-
-import OOPS from '../../assets/error-banner-oops.png';
-
-export const TRACKING_LABELS = {
-  SERVICE_NOT_FOUND: 'service_not_found',
-  UNAUTHORIZED: 'unauthorized',
-  PAGE_LOAD: 'error_during_page_loading',
-};
-
-interface ErrorMessage {
-  message: string;
-  status: number;
-  detail: any;
-}
+import {
+  ErrorMessage,
+  TRACKING_LABELS,
+} from '@ovhcloud/manager-components/src/components/';
+import { ErrorBanner } from '@ovhcloud/manager-components';
 
 interface ErrorObject {
   [key: string]: any;
@@ -37,8 +20,7 @@ function getTrackingTypology(error: ErrorMessage) {
   return TRACKING_LABELS.PAGE_LOAD;
 }
 
-const ErrorBanner: React.FC<ErrorObject> = ({ error }) => {
-  const { t } = useTranslation('{{appName}}/error');
+const Errors: React.FC<ErrorObject> = ({ error }) => {
   const navigate = useNavigate();
   const location = useLocation();
   const shell = useShell();
@@ -46,11 +28,9 @@ const ErrorBanner: React.FC<ErrorObject> = ({ error }) => {
   const env = environment.getEnvironment();
 
   useEffect(() => {
-    tracking.init(false);
     env.then((response) => {
       const { applicationName } = response;
       const name = `errors::${getTrackingTypology(error)}::${applicationName}`;
-      console.info('lance le track page navigation error');
       tracking.trackPage({
         name,
         level2: '81',
@@ -61,52 +41,11 @@ const ErrorBanner: React.FC<ErrorObject> = ({ error }) => {
   }, []);
 
   return (
-    <div className="mx-auto w-full max-w-[600px] grid h-full overflow-hidden p-5">
-      <div className="w-full">
-        <img src={OOPS} alt="OOPS" />
-      </div>
-
-      <div className="py-2">
-        <OsdsText size={ODS_TEXT_SIZE._600} level={ODS_TEXT_LEVEL.heading}>
-          {t('manager_error_page_title')}
-        </OsdsText>
-      </div>
-
-      <div>
-        <OsdsMessage
-          color={ODS_THEME_COLOR_INTENT.error}
-          type={ODS_MESSAGE_TYPE.error}
-        >
-          <div>
-            {t('manager_error_page_default')} <br />
-            {error?.data?.message && <strong>{error.data.message}</strong>}
-            {error?.headers['x-ovh-queryid'] && (
-              <p>
-                {t('manager_error_page_detail_code')}
-                {error.headers['x-ovh-queryid']}
-              </p>
-            )}
-          </div>
-        </OsdsMessage>
-      </div>
-
-      <div className="text-right overflow-hidden py-2">
-        <OsdsButton
-          color={ODS_THEME_COLOR_INTENT.primary}
-          variant={ODS_BUTTON_VARIANT.ghost}
-          onClick={() => navigate('/', { replace: true })}
-        >
-          {t('manager_error_page_action_home_label')}
-        </OsdsButton>
-        <OsdsButton
-          color={ODS_THEME_COLOR_INTENT.primary}
-          variant={ODS_BUTTON_VARIANT.flat}
-          onClick={() => navigate(location.pathname, { replace: true })}
-        >
-          {t('manager_error_page_action_reload_label')}
-        </OsdsButton>
-      </div>
-    </div>
+    <ErrorBanner
+      error={error}
+      onReloadPage={() => navigate(location.pathname, { replace: true })}
+      onRedirectHome={() => navigate('/', { replace: true })}
+    />
   );
 };