+++ packages/manager/apps/nasha/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.20](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.19...@ovh-ux/manager-nasha-app@3.0.20) (2024-02-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
+## [3.0.19](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.18...@ovh-ux/manager-nasha-app@3.0.19) (2024-01-31)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
+## [3.0.18](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.17...@ovh-ux/manager-nasha-app@3.0.18) (2024-01-26)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
+## [3.0.17](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.16...@ovh-ux/manager-nasha-app@3.0.17) (2024-01-25)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
 ## [3.0.16](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.15...@ovh-ux/manager-nasha-app@3.0.16) (2024-01-16)
 
 **Note:** Version bump only for package @ovh-ux/manager-nasha-app