+++ packages/manager/apps/web/CHANGELOG.md
@@ -3,6 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [18.13.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.12.0...@ovh-ux/manager-web@18.13.0) (2024-02-07)
+
+
+### Features
+
+* **web:** remove extra data in call to http ([#10857](https://github.com/ovh/manager/issues/10857)) ([fdecde7](https://github.com/ovh/manager/commit/fdecde777ad20a6a7fe5c6d06201e346a1e18368))
+
+
+
+
+
+# [18.12.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.11.1...@ovh-ux/manager-web@18.12.0) (2024-02-05)
+
+
+### Features
+
+* **web:** add cdn ip information ([#10742](https://github.com/ovh/manager/issues/10742)) ([e5462b7](https://github.com/ovh/manager/commit/e5462b7c7c4d61423162cbef97459207ddaad310))
+* **web:** split dns and domain api calls with two tabs ([#10405](https://github.com/ovh/manager/issues/10405)) ([fc57a4d](https://github.com/ovh/manager/commit/fc57a4d6e9c5599247671385a29f6a6188117da3))
+
+
+
+
+
+## [18.11.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.11.0...@ovh-ux/manager-web@18.11.1) (2024-01-31)
+
+**Note:** Version bump only for package @ovh-ux/manager-web
+
+
+
+
+
+# [18.11.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.10.6...@ovh-ux/manager-web@18.11.0) (2024-01-29)
+
+
+### Features
+
+* **dns:** remove hint about real target with dot ([#10674](https://github.com/ovh/manager/issues/10674)) ([465fd46](https://github.com/ovh/manager/commit/465fd4693169d142f89bfc04f362e592cdfc431c))
+
+
+### Reverts
+
+* Revert "feat(web): add new hosting offer sticker info" (#10705) ([45b23e5](https://github.com/ovh/manager/commit/45b23e526222a36423902203a387b013e13bab61)), closes [#10705](https://github.com/ovh/manager/issues/10705)
+
+
+
+
+
+## [18.10.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.10.5...@ovh-ux/manager-web@18.10.6) (2024-01-26)
+
+**Note:** Version bump only for package @ovh-ux/manager-web
+
+
+
+
+
+## [18.10.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.10.4...@ovh-ux/manager-web@18.10.5) (2024-01-25)
+
+**Note:** Version bump only for package @ovh-ux/manager-web
+
+
+
+
+
 ## [18.10.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.10.3...@ovh-ux/manager-web@18.10.4) (2024-01-22)